requirements.txt
streamlit
scikit-learn
pandas
numpy
joblib
pytesseract
Pillow
requests
